Make messages collapsible

+++ src/api.rs
@@ -122,7 +122,7 @@ async fn list_queues(State(state): State<AppState>) -> Result<Json<Vec<QueueSumm
             QueueSummary {
                 queue_id: local_queue.map(|x| x.id),
                 name: remote_queue.name,
-                message_count_in_rmq: remote_queue.message_count,
+                message_count_in_rmq: remote_queue.message_count as u32,
                 message_count_in_db: local_queue.map(|x| x.message_count).unwrap_or(0),
                 exclusive: remote_queue.exclusive,
             }
